requirements.txt
deepface>=0.0.79
opencv-python>=4.5.5
pandas>=1.3.5
numpy>=1.21.0
matplotlib>=3.5.1
pytz>=2021.1
tensorflow-gpu>=2.8.0
pygame>=2.1.2
gradio>=3.50.2
retina-face>=0.0.13
